My Python Code
This is the code I was planning to turn in, but in the end, my partner's code was more straightforward and had fewer errors. I was too focused on the extra credit portion, but at the end of the day, it still turned out to be different from what I wanted it to be, and I still needed four more prints for the assignment. It is my fault because I was over-focused on creating the pseudo portion instead of the Python code. I wanted to turn in something that I knew was mine but, at the same time, not just ride on the back of my partner.